Use flask.g to store an instance of MPDClient
Connect to MPD on every request using the server details in session, if
available. Redirect GET requests on '/connect' to '/' if the connection
succeeded.

Replace the decorator @require_mpd with two new ones:
 *  @mpdclient
 *  @redirect_on_errors

@mpdclient:
Provides the decorated function with a connected instance of MPDClient
via flask.g.client and raises an exception if the connection could not
be established or the instance returned a command error.

@redirect_on_errors:
Excepts the exceptions raised by @mpdclient and redirects accordingly.